Alles met REST Bulma.

+++ doos-web/src/main/java/eu/debooy/doos/controller/I18nCodeController.java
@@ -31,12 +31,12 @@ import eu.debooy.doosutils.errorhandling.exception.ObjectNotFoundException;
 import eu.debooy.doosutils.errorhandling.exception.base.DoosRuntimeException;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
-import java.util.ArrayList;
-import java.util.Collection;
 import java.util.Properties;
 import javax.enterprise.context.SessionScoped;
+import javax.faces.context.FacesContext;
 import javax.inject.Named;
 import org.apache.commons.io.FilenameUtils;
+import org.json.simple.JSONArray;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -51,37 +51,70 @@ public class I18nCodeController extends Doos {
   private static final  Logger  LOGGER            =
       LoggerFactory.getLogger(I18nCodeController.class);
 
+  private static final  String  DTIT_CREATE   =
+      "doos.titel.i18nCodeTekst.create";
+  private static final  String  DTIT_UPDATE   =
+      "doos.titel.i18nCodeTekst.update";
+  private static final  String  TIT_CREATE    = "doos.titel.i18nCode.create";
+  private static final  String  TIT_RETRIEVE  = "doos.titel.i18nCode.retrieve";
+  private static final  String  TIT_UPDATE    = "doos.titel.i18nCode.update";
+  private static final  String  TIT_UPLOAD    = "doos.titel.i18nCode.upload";
+
   private I18nCode      i18nCode;
   private I18nCodeDto   i18nCodeDto;
   private I18nCodeTekst i18nCodeTekst;
   private Upload        upload;
 
   public void batch() {
+    if (!isUser()) {
+      addError(ComponentsConstants.GEENRECHTEN);
+      return;
+    }
+
     upload  = new Upload();
-    setSubTitel("doos.titel.i18ncode.upload");
+
+    setSubTitel(getTekst(TIT_UPLOAD));
     redirect(I18NUPLOAD_REDIRECT);
   }
 
   public void create() {
+    if (!isUser()) {
+      addError(ComponentsConstants.GEENRECHTEN);
+      return;
+    }
+
     i18nCode    = new I18nCode();
     i18nCodeDto = new I18nCodeDto();
     setAktie(PersistenceConstants.CREATE);
-    setSubTitel("doos.titel.i18ncode.create");
+    setSubTitel(getTekst(TIT_CREATE));
     redirect(I18NCODE_REDIRECT);
   }
 
-  public void createI18nCodeTekst() {
+  public void createDetail() {
+    if (!isUser()) {
+      addError(ComponentsConstants.GEENRECHTEN);
+      return;
+    }
+
     i18nCodeTekst = new I18nCodeTekst();
     i18nCodeTekst.setTaalKode(getGebruikersTaal());
     setDetailAktie(PersistenceConstants.CREATE);
-    setDetailSubTitel("doos.titel.i18ncodetekst.create");
+    setDetailSubTitel(getTekst(DTIT_CREATE));
     redirect(I18NCODETEKST_REDIRECT);
   }
 
-  public void delete(Long codeId, String code) {
+  public void delete() {
+    if (!isUser()) {
+      addError(ComponentsConstants.GEENRECHTEN);
+      return;
+    }
+
+    var codeId  = i18nCode.getCodeId();
+    var code    = i18nCode.getCode();
     try {
       getI18nCodeService().delete(codeId);
       addInfo(PersistenceConstants.DELETED, code);
+      redirect(I18NCODES_REDIRECT);
     } catch (ObjectNotFoundException e) {
       addError(PersistenceConstants.NOTFOUND, code);
     } catch (DoosRuntimeException e) {
@@ -90,11 +123,18 @@ public class I18nCodeController extends Doos {
     }
   }
 
-  public void deleteI18nCodeTekst(String taalKode) {
+  public void deleteDetail() {
+    if (!isUser()) {
+      addError(ComponentsConstants.GEENRECHTEN);
+      return;
+    }
+
+    var taalKode  = i18nCodeTekst.getTaalKode();
     try {
       i18nCodeDto.removeTekst(taalKode);
       getI18nCodeService().save(i18nCodeDto);
       addInfo(PersistenceConstants.DELETED, "'" + taalKode + "'");
+      redirect(I18NCODE_REDIRECT);
     } catch (ObjectNotFoundException e) {
       addError(PersistenceConstants.NOTFOUND, taalKode);
     } catch (DoosRuntimeException e) {
@@ -107,20 +147,14 @@ public class I18nCodeController extends Doos {
     return i18nCode;
   }
 
-  public Collection<I18nCode> getI18nCodes() {
-    return getI18nCodeService().query();
-  }
-
   public I18nCodeTekst getI18nCodeTekst() {
     return i18nCodeTekst;
   }
 
-  public Collection<I18nCodeTekst> getI18nCodeTeksten() {
-    Collection<I18nCodeTekst> teksten = new ArrayList<>();
+  public JSONArray getI18nCodeTeksten() {
+    var teksten = new JSONArray();
 
-    i18nCodeDto.getTeksten()
-               .forEach(i18nCodeTekstDto ->
-                            teksten.add(new I18nCodeTekst(i18nCodeTekstDto)));
+    i18nCodeDto.getTeksten().forEach(rij -> teksten.add(rij.toJSON()));
 
     return teksten;
   }
@@ -129,22 +163,60 @@ public class I18nCodeController extends Doos {
     return upload;
   }
 
-  public void retrieve(Long codeId) {
+  public void retrieve() {
+     if (!isUser() && !isView()) {
+      addError(ComponentsConstants.GEENRECHTEN);
+      return;
+    }
+
+    var ec      = FacesContext.getCurrentInstance().getExternalContext();
+
+    if (!ec.getRequestParameterMap().containsKey(I18nCodeDto.COL_CODEID)) {
+      addError(ComponentsConstants.GEENPARAMETER, I18nCodeDto.COL_CODEID);
+      return;
+    }
+
+    var codeId  =
+        Long.valueOf(ec.getRequestParameterMap()
+                       .get(I18nCodeDto.COL_CODEID));
+
     i18nCodeDto = getI18nCodeService().i18nCode(codeId);
     i18nCode    = new I18nCode(i18nCodeDto);
     setAktie(PersistenceConstants.RETRIEVE);
-    setSubTitel(i18nCode.getCode());
+    setSubTitel(getTekst(TIT_RETRIEVE));
     redirect(I18NCODE_REDIRECT);
   }
 
-  public void retrieveI18nCodeTekst(String taalKode) {
-    i18nCodeTekst = new I18nCodeTekst(i18nCodeDto.getTekst(taalKode));
+  public void retrieveDetail() {
+     if (!isUser()) {
+      addError(ComponentsConstants.GEENRECHTEN);
+      return;
+    }
+
+    var ec  = FacesContext.getCurrentInstance().getExternalContext();
+
+    if (!ec.getRequestParameterMap()
+           .containsKey(I18nCodeTekstDto.COL_TAALKODE)) {
+      addError(ComponentsConstants.GEENPARAMETER,
+               I18nCodeTekstDto.COL_TAALKODE);
+      return;
+    }
+
+    i18nCodeTekst =
+        new I18nCodeTekst(
+            i18nCodeDto.getTekst(ec.getRequestParameterMap()
+                                   .get(I18nCodeTekstDto.COL_TAALKODE)));
     setDetailAktie(PersistenceConstants.RETRIEVE);
     setDetailSubTitel(i18nCode.getCode());
     redirect(I18NCODETEKST_REDIRECT);
   }
 
   public void save() {
+     if (!isUser()) {
+      addError(ComponentsConstants.GEENRECHTEN);
+      return;
+    }
+
     var messages  = I18nCodeValidator.valideer(i18nCode);
     if (!messages.isEmpty()) {
       addMessage(messages);
@@ -178,7 +250,12 @@ public class I18nCodeController extends Doos {
     }
   }
 
-  public void saveI18nCodeTekst() {
+  public void saveDetail() {
+     if (!isUser()) {
+      addError(ComponentsConstants.GEENRECHTEN);
+      return;
+    }
+
     var messages  = I18nCodeTekstValidator.valideer(i18nCodeTekst);
     if (!messages.isEmpty()) {
       addMessage(messages);
@@ -219,24 +296,32 @@ public class I18nCodeController extends Doos {
     }
   }
 
-  public void update(Long codeId) {
-    i18nCodeDto = getI18nCodeService().i18nCode(codeId);
-    i18nCode    = new I18nCode(i18nCodeDto);
+  public void update() {
+     if (!isUser()) {
+      addError(ComponentsConstants.GEENRECHTEN);
+      return;
+    }
+
     setAktie(PersistenceConstants.UPDATE);
-    setSubTitel("doos.titel.i18ncode.update");
-    redirect(I18NCODE_REDIRECT);
+    setSubTitel(getTekst(TIT_UPDATE));
   }
 
-  public void updateI18nCodeTekst(String taalKode) {
-    i18nCodeTekst =
-        new I18nCodeTekst(getI18nCodeService()
-                .i18nCodeTekst(i18nCode.getCodeId(), taalKode));
+  public void updateDetail() {
+     if (!isUser()) {
+      addError(ComponentsConstants.GEENRECHTEN);
+      return;
+    }
+
     setDetailAktie(PersistenceConstants.UPDATE);
-    setDetailSubTitel("doos.titel.i18ncodetekst.update");
-    redirect(I18NCODETEKST_REDIRECT);
+    setDetailSubTitel(getTekst(DTIT_UPDATE));
   }
 
   public void uploading() {
+     if (!isUser()) {
+      addError(ComponentsConstants.GEENRECHTEN);
+      return;
+    }
+
     var bestand = upload.getBestand();
     if (null == bestand) {
       addError("errors.nofile");
